// frontend/src/lib/openapi.d.ts
export interface components {
    schemas: {
// ...
        FolderResponse: {
            /** Id */
            id: string;
            /** Database Id */
            database_id: string;
            /** Name */
            name: string;
            /** Position */
            position: number;
            /** Created At */
            created_at: string;
            /** Updated At */
            updated_at: string;
        };
        /** FolderCreateRequest */
        FolderCreateRequest: {
            /** Database Id */
            database_id: string;
            /** Name */
            name: string;
            /** Position */
            position?: number | null;
        };
        /** FolderUpdateRequest */
        FolderUpdateRequest: {
            /** Name */
            name?: string | null;
            /** Position */
            position?: number | null;
        };
        /** WorksheetResponse */
        WorksheetResponse: {
            /** Id */
            id: string;
            /** Database Id */
            database_id: string;
            /** Name */
            name: string;
            /** Content */
            content: string;
            /** Folder Id */
            folder_id: string | null;
            /** Position */
            position: number;
            /** Created At */
            created_at: string;
            /** Updated At */
            updated_at: string;
        };
        /** WorksheetCreateRequest */
        WorksheetCreateRequest: {
            /** Database Id */
            database_id: string;
            /** Name */
            name: string;
            /**
             * Content
             * @default
             */
            content: string;
            /** Folder Id */
            folder_id?: string | null;
            /** Position */
            position?: number | null;
        };
        /** WorksheetUpdateRequest */
        WorksheetUpdateRequest: {
            /** Name */
            name?: string | null;
            /** Content */
            content?: string | null;
            /** Folder Id */
            folder_id?: string | null;
            /** Position */
            position?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}